+++ CHANGELOG.md
@@ -7,3 +7,5 @@ Unreleased
 --------------------------------
 - adds basic changelog
 - bunch of small tx details UI fixes
+- adds Nomad section on Bridge UI, updates Connext copy
+- adjusts time remaining logic